feat(test): refactor existing tests
